% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_plotting_utils.R
\name{add_events}
\alias{add_events}
\title{Add Event Ticks to Plot}
\usage{
add_events(H_t, col = "#BF616A")
}
\arguments{
\item{H_t}{Numeric vector of event timestamps.}

\item{col}{Hex code for the tick colour. Default is Nord Aurora Red.}
}
\description{
Adds flat-ended rug ticks representing event times.
}
