% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{add_intensity}
\alias{add_intensity}
\title{Add Intensity Line to Plot}
\usage{
add_intensity(t, intensities, col = "#5E81AC")
}
\arguments{
\item{t}{Numeric vector of time points (the grid).}

\item{intensities}{Numeric vector of intensity values.}

\item{col}{Hex code for the line colour. Default is Nord Frost Blue.}
}
\description{
Layers a muted blue intensity curve onto an existing plot.
}
