% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-hp-mle.R
\name{exp_hp_compensator}
\alias{exp_hp_compensator}
\title{Exponential Hawkes Compensator}
\usage{
exp_hp_compensator(theta, H_t, t)
}
\arguments{
\item{theta}{Numeric vector of parameters \eqn{c(\lambda, \alpha, \beta)}.}

\item{H_t}{Numeric vector of event timestamps.}

\item{t}{Numeric value. The time at which to evaluate the compensator
(typically T_max).}
}
\description{
Calculates the integrated intensity \eqn{\Lambda(t)} over the
interval \eqn{[0, t]}.
}
