% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-hp-mle.R
\name{exp_hp_fit}
\alias{exp_hp_fit}
\title{Fit Hawkes Process via MLE}
\usage{
exp_hp_fit(
  object,
  T_max = NULL,
  lik_method = c("fast", "slow"),
  init = c(0.1, 0.5, 1)
)
}
\arguments{
\item{object}{Either a numeric vector of event times or a 'point_process_sim'
object.}

\item{T_max}{Numeric value. Required if 'object' is a numeric vector.}

\item{lik_method}{Character string, either "fast" (recursive) or "slow"
(direct).}

\item{init}{Numeric vector. Initial guesses for the parameters.}
}
\value{
An object of class 'hawkes_fit'.
}
\description{
Estimates parameters \eqn{\lambda, \alpha, \beta} using Maximum
Likelihood.
}
