% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-hp-mle.R
\name{exp_hp_intensity}
\alias{exp_hp_intensity}
\title{Exponential Hawkes Process Intensity}
\usage{
exp_hp_intensity(t, H_t, theta)
}
\arguments{
\item{t}{Numeric value. The time at which to evaluate the intensity.}

\item{H_t}{Numeric vector of event timestamps occurring before \eqn{t}.}

\item{theta}{Numeric vector of parameters \eqn{c(\lambda, \alpha, \beta)}.}
}
\description{
Calculates the conditional intensity \eqn{\lambda^*(t)} at time
\eqn{t}.
}
