% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-hp-mle.R
\name{exp_hp_loglik}
\alias{exp_hp_loglik}
\title{Hawkes Log-Likelihood (Exponential)}
\usage{
exp_hp_loglik(theta, H_t, T_max, lik_method = c("fast", "slow"))
}
\arguments{
\item{theta}{Numeric vector of parameters \eqn{c(\lambda, \alpha, \beta)}.}

\item{H_t}{Numeric vector of event timestamps.}

\item{T_max}{Numeric value. The end of the observation window.}

\item{lik_method}{Character string, either "fast" (recursive) or "slow"
(direct).}
}
\description{
Calculates the log-likelihood of a Hawkes process given a
history.
}
