% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-hp-mle.R
\name{plot.hawkes_fit}
\alias{plot.hawkes_fit}
\title{Plot a Fitted Hawkes Process}
\usage{
\method{plot}{hawkes_fit}(x, n_grid = 1000, ...)
}
\arguments{
\item{x}{An object of class 'hawkes_fit'.}

\item{n_grid}{Integer. The number of points used to calculate the
intensity curve. Default is 1000.}

\item{...}{Additional arguments passed to the plot.}
}
\description{
S3 method for the 'hawkes_fit' class. Plots the estimated
conditional intensity function against the observed event times.
}
