% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-hp-mle.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Q-Q Plot for Hawkes Residuals}
\usage{
plot_residuals(object)
}
\arguments{
\item{object}{An object of class 'hawkes_fit'.}
}
\description{
Checks goodness-of-fit by comparing transformed inter-arrival
times to a theoretical Exponential(1) distribution.
}
