% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp-hp-mle.R
\name{residuals.hawkes_fit}
\alias{residuals.hawkes_fit}
\title{Calculate Transformed Residuals}
\usage{
\method{residuals}{hawkes_fit}(object, ...)
}
\arguments{
\item{object}{An object of class 'hawkes_fit'.}

\item{...}{Additional arguments.}
}
\description{
Calculates the Time-scale transformed residuals for a Hawkes fit
using Thm's 9.1 and 9.2 from Laub, Taimre, Pollett (2021) by converting the
Hawkes process into a unit rate Poisson process by taking the value of the
compensator at each point as the new transformed process.
}
