% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-functions.R
\name{sim_hp}
\alias{sim_hp}
\title{Simulation of a Hawkes process by thinning}
\usage{
sim_hp(T_max, lambda, mu)
}
\arguments{
\item{T_max}{A non-negative numeric value - the end of the interval
\eqn{[0,T_{max}]}.}

\item{lambda}{A non-negative numeric value - the background arrival
component \eqn{\lambda} of the Hawkes conditional intensity}

\item{mu}{A non-negative, non-increasing function - the excitation
function/kernel \eqn{\mu(\cdot)} of the Hawkes conditional intensity.}
}
\value{
An object of class \code{point_process_sim}. A list containing:
\itemize{
\item \code{events}: A sorted numeric vector of arrival times.
\item \code{T_max}: The end of the simulation window.
\item \code{params}: A list of the parameters used for simulation.
\item \code{n}: The total number of events.
}
}
\description{
Simulation of a Hawkes process using Ogata's modified thinning algorithm,
following Chapter 4 (Algorithm 2) of Laub, Taimre, and Pollett (2021).
}
\details{
This method is similar to the thinning method for an inhomogeneous Poisson
process. However, since the Hawkes intensity depends on the event history, a
global upper bound for the conditional intensity is typically not available.
Instead, we assume that the excitation function \eqn{\mu(\cdot)} is
non-increasing. This ensures that, between arrivals, the conditional
intensity is non-increasing in time. As a result, the current intensity can
be used as a valid local upper bound for the next candidate arrival, which is
updated after each event.
}
\examples{
# simulate a Hawkes process with background rate 1 and Exp(2) kernel
t <- sim_hp(T_max=5, lambda = 1, mu = function(t) {2*exp(-2*t)})
}
