% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-functions.R
\name{sim_hp_clus}
\alias{sim_hp_clus}
\title{Simulation of a Hawkes process by clusters}
\usage{
sim_hp_clus(T_max, lambda, eta, family, ...)
}
\arguments{
\item{T_max}{A non-negative numeric value - the end of the interval
\eqn{[0,T_{max}]}.}

\item{lambda}{A non-negative numeric value - the background arrival
component \eqn{\lambda} of the Hawkes conditional intensity}

\item{eta}{A non-negative numeric value - the branching ratio of the Hawkes
process}

\item{family}{A character string "family" which corresponds to a distribution
with a random generation function "rfamily".}

\item{...}{Additional arguments passed to the random generation function.}
}
\value{
An object of class \code{point_process_sim}. A list containing:
\itemize{
\item \code{events}: A sorted numeric vector of arrival times.
\item \code{T_max}: The end of the simulation window.
\item \code{params}: A list of the parameters used for simulation.
\item \code{n}: The total number of events.
\item \code{history}: A dataframe containing the simulation with
additional fields to rebuild the history of immigrant-offspring branches.
}
}
\description{
Simulation of a Hawkes process by clusters i.e. using the immigrant-birth
representation of a Hawkes process. It follows Chapter 4 (Algorithm 3) of
Laub, Taimre, and Pollett (2021). For simplicity, we factor the usual
Hawkes kernel as \eqn{\mu(t) = \eta g(t)} where
\eqn{\eta = \int_0^{\infty} \mu(u) du} is the branching ratio, and \eqn{g(t)}
is a probability density function on \eqn{\mathbb{R}^+}.
}
\details{
Simulation via an immigrant-birth process proceeds in two stages. The first
stage consists of randomly generating immigrant arrivals according to a
Poisson process with intensity \eqn{\lambda} on the interval
\eqn{[0, T_{max}]}.
The second stage then proceeds by randomly generating
\eqn{\mathrm{Poisson}(\eta)} first-generation offspring for each immigrant
generated, where conditional on an immigrant arriving at time \eqn{s}, the
waiting times of its offspring are independent draws from the density
\eqn{g(t)}. Thus, first-generation offspring arrival times are given by
\deqn{s + E_j, \quad E_j \sim g(·)}
Each first-generation offspring in turn generates further descendants
according to the same procedure, which produces the characteristic branching
structure. The recursion continues until no further offspring are generated
within \eqn{[0, T_{max}]}.
}
