% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_plotting_utils.R
\name{plot.point_process_sim}
\alias{plot.point_process_sim}
\title{Plot a Point Process Simulation}
\usage{
\method{plot}{point_process_sim}(x, ...)
}
\arguments{
\item{x}{An object of class 'point_process_sim'.}

\item{...}{Additional arguments (ignored for now).}
}
\description{
S3 method for the point_process_sim class.
}
