% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_funcs.R
\name{sim_hp}
\alias{sim_hp}
\title{Simulation of a Hawkes process by thinning}
\usage{
sim_hp(T_max, lambda, mu)
}
\arguments{
\item{T_max}{A non-negative numeric value - the end of the interval
\eqn{[0,{T_max}]}.}

\item{lambda}{A non-negative numeric value - the background arrival
component \eqn{\lambda} of the Hawkes conditional intensity}

\item{mu}{A non-negative, non-increasing function - the excitation
function/kernel \eqn{\mu(\cdot)} of the Hawkes conditional intensity. Must
be a vectorised function}
}
\value{
An object of class \code{point_process_sim}. A list containing:
\itemize{
\item \code{events}: A sorted numeric vector of arrival times.
\item \code{T_max}: The end of the simulation window.
\item \code{params}: A list of the parameters used for simulation.
\item \code{n}: The total number of events.
\item \code{process}: The class of point process simulated
}
}
\description{
Simulation of a Hawkes process using Ogata's modified thinning algorithm,
following Chapter 4 (Algorithm 2) of Laub, Taimre, and Pollett (2021). This
is similar to the thinning method for an inhomogeneous Poisson process, but
in this case we have no a.s. asymptotic bound \eqn{M} for the conditional
intensity \eqn{\lambda^*(t)}. Instead, we restrict the function space to be
only non-increasing functions of \eqn{t}, which allows us to simply use the
value of \code{fn(t_i)} where \eqn{t_i} is the time just after an arrival, which
is updated after each arrival
}
\examples{
# simulate a Hawkes process with background rate 1 and Exp(2) kernel
t <- sim_hp(T_max=5, lambda = 1, mu = function(t) {2*exp(-2*t)})
}
