% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_pp}
\alias{sim_pp}
\title{Simulation of a Poisson process}
\usage{
sim_pp(T, intens, M = NULL)
}
\arguments{
\item{T}{A non-negative numeric value - the end of the interval \eqn{[0,T]}.}

\item{intens}{A non-negative function (or numeric if simulating a homogeneous
Poisson process) representing the intensity function, \eqn{\lambda (t)} of
the Poisson process.}

\item{M}{A non-negative numeric value - upper bound on \code{fun}. Ignored for
a homogeneous Poisson process.}
}
\value{
A vector of simulated arrival times from the process.
}
\description{
Simulation of a (possibly inhomogeneous) Poisson process using the thinning
algorithm in  Laub, Taimre, and Pollett (2021) Chapter 4 (Algorithm 1). It
works by generating a homogeneous Poisson process with rate \code{M} on
\eqn{[0,T]}, and then, in the inhomogeneous case, thinning this process using
the provided intensity function \code{intens}
}
\examples{
# simulate an inhomogeneous Poisson process with Exp(2) intensity
t <- sim_pp(T = 5, intens = function(t) {2*exp(-2*t)}, M = 2)
# Simulate a homogeneous Poisson process with λ = 2
x <- sim_pp(T = 5, intens = 2)
}
